SCFA - finish separate content from article

+++ server/models/blog/articleContent.js
@@ -6,9 +6,10 @@ class ArticleContent {
     this.articleContent = mongoose.model('ArticleContent', ArticleContentSchema);
   }
 
-  createArticleContent(_content, callback) {
+  createArticleContent(_articleId, _content, callback) {
     const articleContentData = {
       content: _content,
+      articleId: _articleId,
     };
     this.articleContent.create(articleContentData, (err, resultContent) => {
       callback(err, resultContent);
@@ -16,7 +17,17 @@ class ArticleContent {
   }
 
   deleteArticleContent(id, callback) {
-    this.articleContent.deleteOne({ _id: id }, (err) => {
+    if (id !== null) {
+      this.articleContent.deleteOne({ _id: id }, (err) => {
+        callback(err);
+      });
+    } else {
+      callback();
+    }
+  }
+
+  deleteArticleContentByArticleId(id, callback) {
+    this.articleContent.deleteOne({ articleId: id }, (err) => {
       callback(err);
     });
   }
@@ -26,6 +37,12 @@ class ArticleContent {
       content: _content,
     };
     this.articleContent.findByIdAndUpdate(id, articleContentData, (err) => {
+      callback(err);
+    });
+  }
+
+  updateArticleContentByArticleId(id, _content, callback) {
+    this.articleContent.updateOne({ articleId: id }, { $set: { content: _content } }, (err) => {
       callback(err);
     });
   }
